+++ app/cbox/blox/PidBlock.h
@@ -24,7 +24,7 @@ public:
         pid(input, output)
     {}
 
-    static const size_t persistedMaxSize(){
+    static const uint8_t persistedMaxSize(){
         static_assert(blox_Pid_Persisted_size < 128, "varint for settings size will be larger than 1 byte");
         return blox_Pid_Persisted_size + 1;
     }
