+++ app/cbox/blox/SensorSetPointPairBlock.h
@@ -23,7 +23,7 @@ public:
         pair(sensor, setpoint)
 {}
 
-    static const size_t persistedMaxSize(){
+    static const uint8_t persistedMaxSize(){
         static_assert(blox_SensorSetPointPair_Persisted_size < 128, "varint for settings size will be larger than 1 byte");
         return blox_SensorSetPointPair_Persisted_size + 1;
     }
